+++ src/Input/Mouse.cs
@@ -47,7 +47,7 @@ namespace Microsoft.Xna.Framework.Input
 		internal static int INTERNAL_BackBufferWidth = GraphicsDeviceManager.DefaultBackBufferWidth;
 		internal static int INTERNAL_BackBufferHeight = GraphicsDeviceManager.DefaultBackBufferHeight;
 
-		internal static int INTERNAL_MouseWheel = 0;
+		internal static float INTERNAL_MouseWheel = 0;
 
 		#endregion
 
@@ -87,7 +87,7 @@ namespace Microsoft.Xna.Framework.Input
 			return new MouseState(
 				x,
 				y,
-				INTERNAL_MouseWheel,
+				(int) INTERNAL_MouseWheel,
 				left,
 				middle,
 				right,
